dynamics of the inverted wheeled pendulum
Typically, the inverted pendulum is described by a mass that connects to a sliding wagon via a stick. Now the wheeled pendulum is more complex: it has a motor and a transmission, adding one more part (now 3: wheels, motor, body) and adding inertia of rotation. This changes the behavior of the system quite a lot. When the robot is tipping over from stand-still, the robot wheels can actually roll in the SAME direction as the rest of the robot. For the simple inverted pendulum, the wagon always moves OPPOSITE. 

Here is a file used to calculate the dynamics of both systems. 

The python package sympy is used to handle symbolic computations. I find it quite useful, as it helps to focus on what to calculate rather than how.